+++ code/src/sixsq/nuvla/server/resources/configuration_template_nuvla.clj
@@ -28,6 +28,7 @@ notifications will not work.
    :smtp-port   465
    :smtp-ssl    true
    :smtp-debug  true
+
    })