+++ code/src/sixsq/nuvla/server/resources/deployment.clj
@@ -6,11 +6,14 @@ a container orchestration engine.
   (:require
     [clojure.string :as str]
     [sixsq.nuvla.auth.acl-resource :as a]
+    [sixsq.nuvla.auth.acl-resource :as acl-resource]
     [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.db.impl :as db]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
+    [sixsq.nuvla.server.resources.credential :as credential]
+    [sixsq.nuvla.server.resources.customer :as customer]
     [sixsq.nuvla.server.resources.deployment.utils :as dep-utils]
     [sixsq.nuvla.server.resources.event.utils :as event-utils]
     [sixsq.nuvla.server.resources.resource-metadata :as md]
@@ -112,14 +115,19 @@ a container orchestration engine.
 (def add-impl (std-crud/add-fn resource-type collection-acl resource-type))
 
 (defn create-deployment
-  [{:keys [base-uri] :as request}]
+  [{:keys [base-uri] {:keys [owner]} :body :as request}]
   (a/throw-cannot-add collection-acl request)
+  (customer/throw-user-hasnt-active-subscription request)
   (let [authn-info      (auth/current-authentication request)
+        is-admin?       (acl-resource/is-admin? authn-info)
+        dep-owner       (if is-admin? (or owner "group/nuvla-admin")
+                                      (auth/current-user-id request))
         deployment      (-> request
                             (dep-utils/create-deployment)
-                            (assoc :resource-type resource-type)
-                            (assoc :state "CREATED")
-                            (assoc :api-endpoint (str/replace-first base-uri #"/api/" "")))
+                            (assoc :resource-type resource-type
+                                   :state "CREATED"
+                                   :api-endpoint (str/replace-first base-uri #"/api/" "")
+                                   :owner dep-owner))
         ;; FIXME: Correct the value passed to the python API.
 
         create-response (add-impl (assoc request :body deployment))
@@ -151,8 +159,22 @@ a container orchestration engine.
 
 
 (defmethod crud/edit resource-type
-  [request]
-  (edit-impl request))
+  [{{:keys [acl parent]} :body {uuid :uuid} :params :as request}]
+  (let [authn-info (auth/current-authentication request)
+        is-user?   (not (acl-resource/is-admin? authn-info))
+        new-acl    (when (and is-user? acl)
+                     (if-let [current-owner (-> (str resource-type "/" uuid)
+                                                (db/retrieve request)
+                                                :owner)]
+                       (assoc acl :owners (-> acl :owners set (conj current-owner) vec))
+                       acl))
+        infra-id   (some-> parent (crud/retrieve-by-id {:nuvla/authn authn-info}) :parent)]
+
+    (edit-impl
+      (cond-> request
+              is-user? (update :body dissoc :owner :infrastructure-service)
+              new-acl (assoc-in [:body :acl] new-acl)
+              infra-id (assoc-in [:body :infrastructure-service] infra-id)))))
 
 
 (defn delete-impl
@@ -220,6 +242,7 @@ a container orchestration engine.
   (-> resource
       (edit-fn)
       (db/edit request)))
+
 
 (defmethod crud/do-action [resource-type "start"]
   [{{uuid :uuid} :params :as request}]